更新版本到1.2.0，升级gradle到6.3

+++ library/src/main/java/com/ftevxk/core/extension/UnitExtension.kt
@@ -59,7 +59,7 @@ private fun <T : Number> transformUnitValue(value: Float, type: T): T {
         is Int -> value.toInt() as T
         is Double -> value.toDouble() as T
         is Long -> value.toLong() as T
-        is Short -> value.toShort() as T
+        is Short -> value.toInt().toShort() as T
         else -> value as T
     }
 }
